Split serialization tests to reduce the object file size (avoid MinGW object to big issue).
