+++ src/mlpack/tests/serialization_test.cpp
@@ -157,191 +157,6 @@ BOOST_AUTO_TEST_CASE(EmptySparseMatrixSerializeTest)
   TestAllArmadilloSerialization(m);
 }
 
-// Now, test mlpack objects.
-BOOST_AUTO_TEST_CASE(DiscreteDistributionTest)
-{
-  // I assume that I am properly saving vectors, so, this should be
-  // straightforward.
-  vec prob;
-  prob.randu(12);
-  std::vector<arma::vec> prob_vector = std::vector<arma::vec>(1, prob);
-  DiscreteDistribution t(prob_vector);
-
-  DiscreteDistribution xmlT, textT, binaryT;
-
-  // Load and save with all serializers.
-  SerializeObjectAll(t, xmlT, textT, binaryT);
-
-  for (size_t i = 0; i < 12; ++i)
-  {
-    vec obs(1);
-    obs[0] = i;
-    const double prob = t.Probability(obs);
-    if (prob == 0.0)
-    {
-      BOOST_REQUIRE_SMALL(xmlT.Probability(obs), 1e-8);
-      BOOST_REQUIRE_SMALL(textT.Probability(obs), 1e-8);
-      BOOST_REQUIRE_SMALL(binaryT.Probability(obs), 1e-8);
-    }
-    else
-    {
-      BOOST_REQUIRE_CLOSE(prob, xmlT.Probability(obs), 1e-8);
-      BOOST_REQUIRE_CLOSE(prob, textT.Probability(obs), 1e-8);
-      BOOST_REQUIRE_CLOSE(prob, binaryT.Probability(obs), 1e-8);
-    }
-  }
-}
-
-BOOST_AUTO_TEST_CASE(GaussianDistributionTest)
-{
-  vec mean(10);
-  mean.randu();
-  // Generate a covariance matrix.
-  mat cov;
-  cov.randu(10, 10);
-  cov = (cov * cov.t());
-
-  GaussianDistribution g(mean, cov);
-  GaussianDistribution xmlG, textG, binaryG;
-
-  SerializeObjectAll(g, xmlG, textG, binaryG);
-
-  BOOST_REQUIRE_EQUAL(g.Dimensionality(), xmlG.Dimensionality());
-  BOOST_REQUIRE_EQUAL(g.Dimensionality(), textG.Dimensionality());
-  BOOST_REQUIRE_EQUAL(g.Dimensionality(), binaryG.Dimensionality());
-
-  // First, check the means.
-  CheckMatrices(g.Mean(), xmlG.Mean(), textG.Mean(), binaryG.Mean());
-
-  // Now, check the covariance.
-  CheckMatrices(g.Covariance(), xmlG.Covariance(), textG.Covariance(),
-      binaryG.Covariance());
-
-  // Lastly, run some observations through and make sure the probability is the
-  // same.  This should test anything cached internally.
-  arma::mat randomObs;
-  randomObs.randu(10, 500);
-
-  for (size_t i = 0; i < 500; ++i)
-  {
-    const double prob = g.Probability(randomObs.unsafe_col(i));
-
-    if (prob == 0.0)
-    {
-      BOOST_REQUIRE_SMALL(xmlG.Probability(randomObs.unsafe_col(i)), 1e-8);
-      BOOST_REQUIRE_SMALL(textG.Probability(randomObs.unsafe_col(i)), 1e-8);
-      BOOST_REQUIRE_SMALL(binaryG.Probability(randomObs.unsafe_col(i)), 1e-8);
-    }
-    else
-    {
-      BOOST_REQUIRE_CLOSE(prob, xmlG.Probability(randomObs.unsafe_col(i)),
-          1e-8);
-      BOOST_REQUIRE_CLOSE(prob, textG.Probability(randomObs.unsafe_col(i)),
-          1e-8);
-      BOOST_REQUIRE_CLOSE(prob, binaryG.Probability(randomObs.unsafe_col(i)),
-          1e-8);
-    }
-  }
-}
-
-BOOST_AUTO_TEST_CASE(LaplaceDistributionTest)
-{
-  vec mean(20);
-  mean.randu();
-
-  LaplaceDistribution l(mean, 2.5);
-  LaplaceDistribution xmlL, textL, binaryL;
-
-  SerializeObjectAll(l, xmlL, textL, binaryL);
-
-  BOOST_REQUIRE_CLOSE(l.Scale(), xmlL.Scale(), 1e-8);
-  BOOST_REQUIRE_CLOSE(l.Scale(), textL.Scale(), 1e-8);
-  BOOST_REQUIRE_CLOSE(l.Scale(), binaryL.Scale(), 1e-8);
-
-  CheckMatrices(l.Mean(), xmlL.Mean(), textL.Mean(), binaryL.Mean());
-}
-
-BOOST_AUTO_TEST_CASE(MahalanobisDistanceTest)
-{
-  MahalanobisDistance<> d;
-  d.Covariance().randu(50, 50);
-
-  MahalanobisDistance<> xmlD, textD, binaryD;
-
-  SerializeObjectAll(d, xmlD, textD, binaryD);
-
-  // Check the covariance matrices.
-  CheckMatrices(d.Covariance(),
-                xmlD.Covariance(),
-                textD.Covariance(),
-                binaryD.Covariance());
-}
-
-BOOST_AUTO_TEST_CASE(LinearRegressionTest)
-{
-  // Generate some random data.
-  mat data;
-  data.randn(15, 800);
-  rowvec responses;
-  responses.randn(800);
-
-  LinearRegression lr(data, responses, 0.05); // Train the model.
-  LinearRegression xmlLr, textLr, binaryLr;
-
-  SerializeObjectAll(lr, xmlLr, textLr, binaryLr);
-
-  BOOST_REQUIRE_CLOSE(lr.Lambda(), xmlLr.Lambda(), 1e-8);
-  BOOST_REQUIRE_CLOSE(lr.Lambda(), textLr.Lambda(), 1e-8);
-  BOOST_REQUIRE_CLOSE(lr.Lambda(), binaryLr.Lambda(), 1e-8);
-
-  CheckMatrices(lr.Parameters(), xmlLr.Parameters(), textLr.Parameters(),
-      binaryLr.Parameters());
-}
-
-BOOST_AUTO_TEST_CASE(RegressionDistributionTest)
-{
-  // Generate some random data.
-  mat data;
-  data.randn(15, 800);
-  rowvec responses;
-  responses.randn(800);
-
-  RegressionDistribution rd(data, responses);
-  RegressionDistribution xmlRd, textRd, binaryRd;
-
-  // Okay, now save it and load it.
-  SerializeObjectAll(rd, xmlRd, textRd, binaryRd);
-
-  // Check the gaussian distribution.
-  CheckMatrices(rd.Err().Mean(),
-                xmlRd.Err().Mean(),
-                textRd.Err().Mean(),
-                binaryRd.Err().Mean());
-  CheckMatrices(rd.Err().Covariance(),
-                xmlRd.Err().Covariance(),
-                textRd.Err().Covariance(),
-                binaryRd.Err().Covariance());
-
-  // Check the regression function.
-  if (rd.Rf().Lambda() == 0.0)
-  {
-    BOOST_REQUIRE_SMALL(xmlRd.Rf().Lambda(), 1e-8);
-    BOOST_REQUIRE_SMALL(textRd.Rf().Lambda(), 1e-8);
-    BOOST_REQUIRE_SMALL(binaryRd.Rf().Lambda(), 1e-8);
-  }
-  else
-  {
-    BOOST_REQUIRE_CLOSE(rd.Rf().Lambda(), xmlRd.Rf().Lambda(), 1e-8);
-    BOOST_REQUIRE_CLOSE(rd.Rf().Lambda(), textRd.Rf().Lambda(), 1e-8);
-    BOOST_REQUIRE_CLOSE(rd.Rf().Lambda(), binaryRd.Rf().Lambda(), 1e-8);
-  }
-
-  CheckMatrices(rd.Rf().Parameters(),
-                xmlRd.Rf().Parameters(),
-                textRd.Rf().Parameters(),
-                binaryRd.Rf().Parameters());
-}
-
 BOOST_AUTO_TEST_CASE(BallBoundTest)
 {
   BallBound<> b(100);
@@ -362,7 +177,7 @@ BOOST_AUTO_TEST_CASE(BallBoundTest)
   BOOST_REQUIRE_CLOSE(b.Radius(), textB.Radius(), 1e-8);
   BOOST_REQUIRE_CLOSE(b.Radius(), binaryB.Radius(), 1e-8);
 
-  // Now check the vectors.
+  // Now check the vectors.ƒ
   CheckMatrices(b.Center(), xmlB.Center(), textB.Center(), binaryB.Center());
 }
 
@@ -1787,62 +1602,6 @@ BOOST_AUTO_TEST_CASE(ssRBMTest)
   CheckMatrices(Rbm.Weight(), RbmXml.Weight());
   CheckMatrices(Rbm.Weight(), RbmText.Weight());
   CheckMatrices(Rbm.Weight(), RbmBinary.Weight());
-}
-
-// General ANN serialization test.
-template<typename LayerType>
-void ANNLayerSerializationTest(LayerType& layer)
-{
-  arma::mat input(5, 100, arma::fill::randu);
-  arma::mat output(5, 100, arma::fill::randu);
-
-  FFN<NegativeLogLikelihood<>, ann::RandomInitialization> model;
-  model.Add<Linear<>>(input.n_rows, 10);
-  model.Add<LayerType>(layer);
-  model.Add<ReLULayer<>>();
-  model.Add<Linear<>>(10, output.n_rows);
-  model.Add<LogSoftMax<>>();
-
-  ens::StandardSGD opt(0.1, 1, 5, -100, false);
-  model.Train(input, output, opt);
-
-  arma::mat originalOutput;
-  model.Predict(input.col(0), originalOutput);
-
-  // Now serialize the model.
-  FFN<NegativeLogLikelihood<>, ann::RandomInitialization> xmlModel, textModel,
-      binaryModel;
-  SerializeObjectAll(model, xmlModel, textModel, binaryModel);
-
-  // Ensure that predictions are the same.
-  arma::mat modelOutput, xmlOutput, textOutput, binaryOutput;
-  model.Predict(input.col(0), modelOutput);
-  xmlModel.Predict(input.col(0), xmlOutput);
-  textModel.Predict(input.col(0), textOutput);
-  binaryModel.Predict(input.col(0), binaryOutput);
-
-  CheckMatrices(originalOutput, modelOutput, 1e-5);
-  CheckMatrices(originalOutput, xmlOutput, 1e-5);
-  CheckMatrices(originalOutput, textOutput, 1e-5);
-  CheckMatrices(originalOutput, binaryOutput, 1e-5);
-}
-
-/**
- * Simple serialization test for batch normalization layer.
- */
-BOOST_AUTO_TEST_CASE(BatchNormSerializationTest)
-{
-  BatchNorm<> layer(10);
-  ANNLayerSerializationTest(layer);
-}
-
-/**
- * Simple serialization test for layer normalization layer.
- */
-BOOST_AUTO_TEST_CASE(LayerNormSerializationTest)
-{
-  LayerNorm<> layer(10);
-  ANNLayerSerializationTest(layer);
 }
 
 BOOST_AUTO_TEST_SUITE_END();
